refactor(seshat): split ui into per-section modules
ui/mod.rs had grown into a 411-line grab-bag (connections, editor, results grid,
schema tree, history, dialog, and shared primitives all in one file). Split it
into focused modules, each owning one section:

  ui/mod.rs      — module wiring + build_ui dispatch (28 lines)
  ui/widgets.rs  — shared builders (button, text_input, muted, caret, indent)
  ui/connections.rs — connections manager + list
  ui/editor.rs   — SQL editor + typed results grid
  ui/schema.rs   — schema-browser tree
  ui/history.rs  — query-history list
  ui/dialog.rs   — new/edit-connection modal
  ui/sidebar.rs  — tabbed navigator

Pure code movement — no behavior change. clippy (host + wasm) + fmt clean;
197 tests pass.
