+++ plugins/seshat/src/ui/sidebar.rs
@@ -7,7 +7,10 @@
 use serde_json::{json, Value};
 
 use crate::state::State;
-use crate::ui::{connections_list, dialog, history_list, schema_panel};
+use crate::ui::connections::connections_list;
+use crate::ui::dialog::dialog;
+use crate::ui::history::history_list;
+use crate::ui::schema::schema_panel;
 use crate::{ICON_HISTORY, ICON_PLUGS_CONNECTED, ICON_PLUS, ICON_TREE_STRUCTURE};
 
 pub(crate) fn build_sidebar(st: &State) -> Value {
